added some secs
