+++ src/controllers/post.ts
@@ -56,21 +56,21 @@ async function createPostHandler(req: Request, res: Response): Promise<any> {
 
         const media = files
             ? await Promise.all(
-                  files.map(async (file) => {
-                      const formData = new FormData();
-                      formData.append("image", fs.createReadStream(file.path));
+                files.map(async (file) => {
+                    const formData = new FormData();
+                    formData.append("image", fs.createReadStream(file.path));
 
-                      const response = await UploadService.upload(formData);
-                      const imageUrl = response.data.data.url;
+                    const response = await UploadService.upload(formData);
+                    const imageUrl = response.data.data.url;
 
-                      fs.unlinkSync(file.path);
+                    fs.unlinkSync(file.path);
 
-                      return {
-                          url: imageUrl,
-                          type: mediaType(file.mimetype),
-                      };
-                  })
-              )
+                    return {
+                        url: imageUrl,
+                        type: mediaType(file.mimetype),
+                    };
+                })
+            )
             : [];
 
         const post = await Post.create({
@@ -219,422 +219,10 @@ async function deletePostHandler(req: Request, res: Response): Promise<any> {
     }
 }
 
-/* Applaud Handlers */
-
-/***
- * @desc Get all applauds in a post
- * @route GET /api/posts/post/:postId/applauds
- */
-async function allApplaudsHandler(req: Request, res: Response): Promise<any> {
-    const { postId } = req.params;
-
-    try {
-        let response: IResponse = {
-            msg: "",
-        };
-
-        const post = await Post.findById(postId).populate("applauds");
-
-        if (!post) {
-            response.msg = "Post Not found";
-            return res.status(404).json(response);
-        }
-
-        response.msg = post.applauds.length === 0 ? "No Applauds Yet!" : "Fetched all users applauds";
-        response.data = post.applauds as unknown as IUser[];
-
-        return res.status(200).json(response);
-    } catch (error) {
-        console.log(error);
-        return res.status(500).json(setErrorDetails("Internal Server Error", error as string));
-    }
-}
-
-/***
- * @desc Applaud or unApplaud in a post
- * @route GET /api/posts/post/:postId/applaud
- */
-async function applaudPostHandler(req: Request, res: Response): Promise<any> {
-    const { postId } = req.params;
-    const userId = req.user?.id;
-
-    try {
-        let response: IResponse = {
-            msg: "",
-        };
-
-        const post = await Post.findById(postId);
-        if (!post) {
-            response.msg = "Post Not found";
-            return res.status(404).json(response);
-        }
-
-        const isLiked: boolean = post?.applauds.includes(userId);
-
-        const updatedPost = await Post.findByIdAndUpdate(postId, isLiked ? { $pull: { applauds: userId } } : { $addToSet: { applauds: userId } }, { new: true });
-
-        response.msg = isLiked ? "UnApplauded a Post" : "Applauded a Post";
-        response.data = updatedPost!;
-        return res.status(200).json(response);
-    } catch (error) {
-        return res.status(500).json(setErrorDetails("Internal Server Error", error as string));
-    }
-}
-
-/* Comment Handler */
-
-/***
- * @desc Get all comments in a post
- * @route GET /api/posts/post/:postId/comments
- */
-async function allCommentsHandler(req: Request, res: Response): Promise<any> {
-    const { postId } = req.params;
-
-    try {
-        let response: IResponse = {
-            msg: "",
-        };
-
-        const post = await Post.findById(postId).populate([{ path: "comments.user" }, { path: "comments.applauds" }, { path: "comments.replies.user" }]);
-        if (!post) {
-            response.msg = "Post Not found";
-            return res.status(404).json(response);
-        }
-
-        response.msg = post.comments.length === 0 ? "No Comments Yet!" : "Fetched all comments";
-        response.data = post.comments;
-
-        return res.status(200).json(response);
-    } catch (error) {
-        console.log(error);
-        return res.status(500).json(setErrorDetails("Internal Server Error", error as string));
-    }
-}
-
-/***
- * @desc Comment in a post
- * @route POST /api/posts/post/:postId/comment
- */
-async function commentPostHandler(req: Request, res: Response): Promise<any> {
-    const { postId } = req.params;
-    const userId = req.user?.id as unknown as Schema.Types.ObjectId;
-    const { text } = req.body;
-
-    try {
-        let response: IResponse = {
-            msg: "",
-        };
-
-        const post = await Post.findByIdAndUpdate(
-            postId,
-            {
-                $push: { comments: { user: userId, text: text } },
-            },
-            { new: true }
-        );
-
-        if (!post) {
-            response.msg = "No Post Found";
-            return res.status(404).json(response);
-        }
-
-        response.msg = "Commented Successfully";
-        response.data = post;
-        return res.status(200).json(response);
-    } catch (error) {
-        return res.status(500).json(setErrorDetails("Internal Server Error", error as string));
-    }
-}
-
-/***
- * @desc Update a comment
- * @route Patch /api/posts/post/:postId/comment/:commentId
- */
-async function updateCommentHandler(req: Request, res: Response): Promise<any> {
-    const { postId, commentId } = req.params;
-    let { text } = req.body;
-
-    try {
-        let response: IResponse = {
-            msg: "",
-        };
-
-        const post = await Post.findById(postId);
-        if (!post) {
-            response.msg = "No Post Found";
-            return res.status(404).json(response);
-        }
-
-        const comment = post.comments.find((comment) => comment._id.toString() === commentId);
-
-        if (!comment) {
-            response.msg = "No Comment Found";
-            return res.status(404).json(response);
-        }
-
-        comment.text = text;
-        await post.save();
-
-        response.msg = "Comment Updated Successfully";
-        response.data = comment;
-        return res.status(200).json(response);
-    } catch (error) {
-        return res.status(500).json(setErrorDetails("Internal Server Error", error as string));
-    }
-}
-
-/***
- * @desc Delete a comment
- * @route Delete /api/posts/post/:postId/comment/:commentId
- */
-async function deleteCommentHandler(req: Request, res: Response): Promise<any> {
-    const { commentId, postId } = req.params;
-
-    try {
-        let response: IResponse = {
-            msg: "",
-        };
-
-        const post = await Post.findById(postId);
-        if (!post) {
-            response.msg = "No Post Found";
-            return res.status(404).json(response);
-        }
-
-        post.comments = post.comments.filter((comment) => comment._id.toString() !== commentId);
-        await post.save();
-
-        response.msg = "Comment Deleted Successfully";
-        response.data = post.comments;
-        return res.status(200).json(response);
-    } catch (error) {
-        return res.status(500).json(setErrorDetails("Internal Server Error", error as string));
-    }
-}
-
-/* Comment Interaction Controllers */
-
-/***
- * @desc Applaud in a post comment
- * @route GET /api/posts/post/comment/like/:commentId
- */
-async function addPostCommentApplaudHandler(req: Request, res: Response): Promise<any> {
-    const userId = req.user?.id;
-    const { commentId, postId } = req.params;
-
-    try {
-        let response: IResponse = {
-            msg: "",
-        };
-
-        const post = await Post.findById(postId);
-        if (!post) {
-            response.msg = "Post Not found";
-            return res.status(404).json(response);
-        }
-
-        const comment = post.comments.find((comment) => comment._id.toString() === commentId);
-        if (!comment) {
-            response.msg = "Comment Not found";
-            return res.status(404).json(response);
-        }
-
-        const isApplauded: boolean = (comment.applauds ?? []).includes(userId);
-        if (isApplauded) {
-            comment.applauds = comment.applauds?.filter((applaud) => applaud.toString() !== userId);
-        } else {
-            comment.applauds?.push(userId);
-        }
-        await post.save();
-
-        response.msg = isApplauded ? "UnApplauded Comment" : "Applauded Comment";
-        response.data = comment;
-        return res.status(200).json(response);
-    } catch (error) {
-        return res.status(500).json(setErrorDetails("Internal Server Error", error as string));
-    }
-}
-
-/***
- * @desc Reply in a post comment
- * @route POST /api/posts/post/:postId/comment/:commentId/reply
- */
-async function addReplyHandler(req: Request, res: Response): Promise<any> {
-    const { text } = req.body;
-    const { postId, commentId } = req.params;
-    const userId = req.user?.id;
-
-    try {
-        let response: IResponse = {
-            msg: "",
-        };
-
-        const post = await Post.findById(postId);
-        if (!post) {
-            response.msg = "No Post Found";
-            return res.status(404).json(response);
-        }
-
-        const comment = post.comments.find((comment) => comment._id.toString() === commentId);
-        if (!comment) {
-            response.msg = "Comment Not found";
-            return res.status(404).json(response);
-        }
-
-        comment.replies = comment.replies || [];
-        comment.replies.push({
-            user: userId,
-            text: text,
-        });
-        await post.save();
-
-        response.msg = "Replied Successfully";
-        response.data = comment;
-        return res.status(200).json(response);
-    } catch (error) {
-        return res.status(500).json(setErrorDetails("Internal Server Error", error as string));
-    }
-}
-
-/**
- * @desc Update a reply
- * @route Patch /api/posts/post/comment/reply/:commentId
- */
-async function updateReplyHandler(req: Request, res: Response): Promise<any> {
-    const { text } = req.body;
-    const { postId, commentId, replyId } = req.params;
-
-    try {
-        let response: IResponse = {
-            msg: "",
-        };
-        const post = await Post.findById(postId);
-        if (!post) {
-            response.msg = "No Post Found";
-            return res.status(404).json(response);
-        }
-
-        const comment = post.comments.find((comment) => comment._id.toString() === commentId);
-        if (!comment) {
-            response.msg = "No Comment Found";
-            return res.status(404).json(response);
-        }
-
-        const reply = comment.replies?.find((reply) => reply._id!.toString() === replyId);
-        if (!reply) {
-            response.msg = "No Reply Found";
-            return res.status(404).json(response);
-        }
-
-        reply.text = text;
-        await post.save();
-
-        response.msg = "Reply Updated Successfully";
-        response.data = comment;
-        return res.status(200).json(response);
-    } catch (error) {
-        return res.status(500).json(setErrorDetails("Internal Server Error", error as string));
-    }
-}
-
-/**
- * @desc Delete a reply
- * @route Patch /api/posts/post/:postId/:commentId/comment/reply/:replyId
- */
-async function deleteReplyHandler(req: Request, res: Response): Promise<any> {
-    const { postId, commentId, replyId } = req.params;
-
-    try {
-        let response: IResponse = {
-            msg: "",
-        };
-
-        const post = await Post.findById(postId);
-        if (!post) {
-            response.msg = "No Post Found";
-            return res.status(404).json(response);
-        }
-
-        const comment = post.comments.find((comment) => comment._id.toString() === commentId);
-        if (!comment) {
-            response.msg = "No Comment Found";
-            return res.status(404).json(response);
-        }
-
-        if (!comment.replies || comment.replies.length === 0) {
-            response.msg = "No Replies Found";
-            return res.status(404).json(response);
-        }
-
-        comment.replies = comment.replies.filter((reply) => reply._id!.toString() !== replyId);
-        await post.save();
-
-        response.msg = "Reply Deleted Successfully";
-        response.data = comment;
-        return res.status(200).json(response);
-    } catch (error) {
-        return res.status(500).json(setErrorDetails("Internal Server Error", error as string));
-    }
-}
-
-/**
- * @desc Add Applaud in a reply
- * @route GET /api/posts/post/:postId/comment/:commentId/reply/:replyId/applaud
- */
-async function addApplaudPostCommentReplyHandler(req: Request, res: Response): Promise<any> {
-    const { postId, commentId, replyId } = req.params;
-    const userId = req.user?.id;
-
-    try {
-        let response: IResponse = {
-            msg: "",
-        };
-
-        const post = await Post.findById(postId);
-        if (!post) {
-            response.msg = "No Post Found";
-            return res.status(404).json(response);
-        }
-
-        const comment = post.comments.find((comment) => comment._id.toString() === commentId);
-        if (!comment) {
-            response.msg = "No Comment Found";
-            return res.status(404).json(response);
-        }
-
-        const reply = (comment.replies ?? []).find((reply) => reply._id!.toString() === replyId);
-        if (!reply) {
-            response.msg = "No Reply Found";
-            return res.status(404).json(response);
-        }
-
-        reply.applauds = reply.applauds ?? [];
-        reply.applauds.push(userId);
-        await post.save();
-
-        response.msg = "Applauded Successfully in the Reply";
-        response.data = comment;
-        return res.status(200).json(response);
-    } catch (error) {
-        return res.status(500).json(setErrorDetails("Internal Server Error", error as string));
-    }
-}
-
 export {
     createPostHandler,
     updatePostHandler,
     getPostsHandler,
     getSpecificPostHandler,
-    applaudPostHandler,
-    allApplaudsHandler,
-    commentPostHandler,
     deletePostHandler,
-    allCommentsHandler,
-    addReplyHandler,
-    addPostCommentApplaudHandler,
-    deleteCommentHandler,
-    updateCommentHandler,
-    updateReplyHandler,
-    deleteReplyHandler,
-    addApplaudPostCommentReplyHandler,
 };
