+++ src/routes/api/index.ts
@@ -16,11 +16,11 @@ router.get("/verify-email", verifyEmailHandler);
 
 router.use(authCheckMiddleware);
 router.use("/user", userRoutes);
-router.use("/posts", postsRoutes);
 router.use("/collections", collectionRoutes);
-router.use("/socials", socialRoutes);
 router.use("/discover", discoverRoutes);
 router.use("/conversation", conversationRoutes);
+router.use("/posts", postsRoutes);
 router.use("/story", storyRoutes);
+router.use("/socials", socialRoutes);
 
 export default router;
